pyproject.toml
[tool.poetry]
name = "srpski"
version = "0.1.0"
description = ""
authors = ["Vladimir Rybakov <vladimir.rybakov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
anyio = "4.3.0"
boto3 = "1.34.64"
botocore = "1.34.64"
cachetools = "5.3.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
colorclass = "2.2.2"
docopt = "0.6.2"
google-api-core = "2.17.1"
google-api-python-client = "2.121.0"
google-auth = "2.28.1"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.0"
googleapis-common-protos = "1.62.0"
greenlet = "3.0.3"
gspread = "6.0.2"
h11 = "0.14.0"
httpcore = "1.0.4"
httplib2 = "0.22.0"
httpx = "0.26.0"
idna = "3.6"
jmespath = "1.0.1"
oauth2client = "4.1.3"
oauthlib = "3.2.2"
packaging = "23.2"
pillow = "10.2.0"
pip-upgrader = "1.4.15"
protobuf = "4.25.3"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pympler = "1.0.1"
pyparsing = "3.1.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-telegram-bot = "20.8"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rsa = "4.9"
s3transfer = "0.10.1"
setuptools = "69.1.1"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.28"
strenum = "0.4.15"
terminaltables = "3.1.10"
typing-extensions = "4.10.0"
uritemplate = "4.1.1"
urllib3 = "2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
